# CONSOLIDATED: Critical Tasks Resource Fixes + Enhanced Error Messages (#415, #416, #417)