… src/utils/enhanced-validation.ts
… src/utils/validation/uuid-validation.ts
… test/e2e/utils/test-data.ts
… test/integration/consolidated-fixes.test.ts
… test/performance/regression.test.ts
… test/unit/handlers/universal/tasks-attribute-discovery.test.ts
… test/unit/handlers/universal/tasks-crud-simulation.test.ts
… test/unit/handlers/universal/tasks-field-mapping.test.ts